Fix libwm/automake dependency issue.
Passing libwm.a instead of -lwm forces wm_driver to be rebuilt if libwm
is modified, which is what needs to happen.  There might be a cleaner
solution, but this is how the other Makefiles work.

Also added types.h to SOURCES so that it gets picked up by make dist.