<decimal_digits>G support for jumping to a line
I wanted the vim style jump to line support,
for example entering the sequence "101G" would jump your cursor to the
101st line.

The way I've implemented this is, well, fairly hackish, but is no more
limited than the 'G' command already is. Specifically,the G command does
not jump to the end line of an arbitrarily long file, it jumps to line
10000000. Because of this, it is reasonable to only support 8 digits of
jump.

Signed-off-by: Michael Sartain <mikesart@gmail.com>